move ArtifactLayoutReference class to internal.io.reference package
